Code for both programs together has been created
Only the social justice section is complete. For financial literacy, the program starts with an input dialog box. However, all of the rest of the questions and inputs are still showing in the command window. We might have to forgo randomization unless we figure out how to code that in graphical form?
A figure (see: imshow(bg) on line 104) displays all of the variable stats at the end of each 'round.'